feat: table naming provider + MANAGE_SESSIONS flag
- TableNaming enum (Pascal/Lower) + TableNames com nomes de tabelas
- DB_TABLE_NAMING env var (pascal | lower)
- MANAGE_SESSIONS flag: desativa Redis session management p/ Go/Node
- Logout aceita config para conditional session invalidation
- Cache usa RedisKeys configurável por naming convention
- Login/Refresh criam sessions condicionalmente

+++ src/models/auth.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub password: Option<String>,
     pub request_password_token: Option<String>,
     pub request_password_expiration: Option<DateTimeWithTimeZone>,
-    pub retries: i32,
+    pub retries: i64,
     pub first_access: bool,
     pub active: bool,
     pub is_deleted: Option<bool>,
